+++ internal/dataplane/kong_client.go
@@ -40,7 +40,9 @@ const (
 	KongConfigurationApplyFailedEventReason = "KongConfigurationApplyFailed"
 )
 
-type ClientsProvider interface {
+// AdminAPIClientsProvider allows fetching the most recent list of Admin API clients of Gateways that
+// we should configure.
+type AdminAPIClientsProvider interface {
 	AllClients() []*adminapi.Client
 	GatewayClients() []*adminapi.Client
 }
@@ -132,7 +134,7 @@ type KongClient struct {
 	SHAs []string
 
 	// clientsProvider allows retrieving the most recent set of clients.
-	clientsProvider ClientsProvider
+	clientsProvider AdminAPIClientsProvider
 }
 
 // NewKongClient provides a new KongClient object after connecting to the
@@ -147,7 +149,7 @@ func NewKongClient(
 	kongConfig sendconfig.Config,
 	eventRecorder record.EventRecorder,
 	dbMode string,
-	clientsProvider ClientsProvider,
+	clientsProvider AdminAPIClientsProvider,
 ) (*KongClient, error) {
 	// build the client object
 	cache := store.NewCacheStores()
@@ -260,7 +262,6 @@ func (c *KongClient) Listeners(ctx context.Context) ([]kong.ProxyListener, []kon
 	c.lock.RLock()
 	for _, cl := range c.clientsProvider.GatewayClients() {
 		cl := cl
-
 		errg.Go(func() error {
 			listeners, streamListeners, err := cl.AdminAPIClient().Listeners(ctx)
 			if err != nil {
@@ -386,7 +387,7 @@ func (c *KongClient) Update(ctx context.Context) error {
 	defer c.lock.Unlock()
 
 	// build the kongstate object from the Kubernetes objects in the storer
-	storer := store.New(*c.cache, c.ingressClass, false, false, false, c.logger)
+	storer := store.New(*c.cache, c.ingressClass, c.logger)
 
 	// initialize a parser
 	c.logger.Debug("parsing kubernetes objects into data-plane configuration")